Merge commit '4e83f94696d1930d10fd3bf4d1c6c6a589297782' into HEAD
"set all usb mode to none" is replaced by "save all usb mode to file"

+++ meta-qti-base/recipes-kernel/linux-msm/linux-msm_5.4.bb
@@ -181,7 +181,9 @@ do_deploy () {
     install -d ${DEPLOYDIR}/build-artifacts/kernel_scripts/scripts
     install -d ${DEPLOYDIR}/build-artifacts/kernel_scripts/usr
     cp  ${STAGING_KERNEL_DIR}/usr/gen_initramfs_list.sh ${DEPLOYDIR}/build-artifacts/kernel_scripts/scripts
-    cp  ${STAGING_KERNEL_BUILDDIR}/usr/gen_init_cpio ${DEPLOYDIR}/build-artifacts/kernel_scripts/usr
+    if [ -f ${STAGING_KERNEL_BUILDDIR}/usr/gen_init_cpio ]; then
+        cp  ${STAGING_KERNEL_BUILDDIR}/usr/gen_init_cpio ${DEPLOYDIR}/build-artifacts/kernel_scripts/usr
+    fi
 
     # Copy Image appended with dtbs to deploydir
     cat ${D}/${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}-${KERNEL_VERSION} ${B}/arch/${ARCH}/boot/dts/vendor/qcom/*.dtb > ${DEPLOYDIR}/${KERNEL_IMAGETYPE}-dtb-${KERNEL_VERSION}
